AddProductForm incapsulated
Also I created OwnerProductInput but it has minor changes in comparing with InputForm
