+++ app/components/ui/Modals/AdminPanel/components/AddProductForm.tsx
@@ -1,17 +1,155 @@
-import { FieldErrors, UseFormRegister } from "react-hook-form"
-import { ProductInput } from "../../../Inputs/Validation"
-import { Button } from "../../.."
+import { useEffect, useState, useRef } from "react"
+import Image from "next/image"
+import axios from "axios"
+import { useForm } from "react-hook-form"
+import supabaseClient from "@/utils/supabaseClient"
+import { Stripe, loadStripe } from "@stripe/stripe-js"
+import { ImageListType } from "react-images-uploading"
+import ImageUploading from "react-images-uploading"
 
-interface AddProductForm {
-  register: UseFormRegister<any>
-  errors: FieldErrors
-  responseMessage: React.ReactNode
-  isLoading: boolean
-}
+import useUserStore from "@/store/user/userStore"
+import { IFormDataAddProduct } from "@/interfaces/IFormDataAddProduct"
+import { ProductInput } from "@/components/ui/Inputs/Validation"
+import { Button } from "@/components/ui/Button"
+import useDragging from "@/hooks/ui/useDragging"
+import { twMerge } from "tailwind-merge"
 
-export default function AddProductForm({ register, errors, responseMessage, isLoading }: AddProductForm) {
+export default function AddProductForm() {
+  const userStore = useUserStore()
+  const { isDraggingg } = useDragging()
+
+  const [isLoading, setIsLoading] = useState(false)
+  const [stripe, setStripe] = useState<Stripe | null>(null)
+  const [responseMessage, setResponseMessage] = useState<React.ReactNode>(<p></p>)
+  const [images, setImages] = useState<ImageListType>([])
+
+  const dragZone = useRef<HTMLButtonElement | null>(null)
+
+  useEffect(() => {
+    const initializeStripe = async () => {
+      const stripeInstance = await loadStripe(process.env.NEXT_PUBLIC_STRIPE_PUBLIC)
+      setStripe(stripeInstance)
+    }
+    initializeStripe()
+  }, [])
+
+  async function createProduct(images: ImageListType, title: string, subTitle: string, price: number, onStock: number) {
+    setIsLoading(true)
+    try {
+      //Check images length and is stripe mounted
+      if (images.length > 0 && stripe) {
+        const imagesArray = await Promise.all(
+          images.map(async image => {
+            if (image?.file && userStore.userId) {
+              const { data, error } = await supabaseClient.storage
+                .from("public")
+                .upload(`${userStore.userId}/${image.file.name}`, image.file, { upsert: true })
+              if (error) throw error
+
+              const response = supabaseClient.storage.from("public").getPublicUrl(data.path)
+              return response.data.publicUrl
+            }
+          }),
+        )
+        //create product
+        const priceResponse = await axios.post("/api/products", {
+          images: imagesArray,
+          title: title,
+          subTitle: subTitle,
+          price: price,
+        })
+
+        const updatedUserResponse = await supabaseClient
+          .from("products")
+          .insert({
+            id: priceResponse.data.id,
+            owner_id: userStore.userId,
+            title: title,
+            sub_title: subTitle,
+            price: price,
+            on_stock: onStock,
+            img_url: imagesArray as string[],
+          })
+          .eq("user_id", userStore.userId)
+        if (updatedUserResponse.error) throw updatedUserResponse.error
+        displayResponseMessage(<p className="text-success">Product added</p>)
+      } else {
+        displayResponseMessage(<p className="text-danger">Upload the image</p>)
+      }
+    } catch (error) {
+      console.log(94, "error - ", error)
+    } finally {
+      setIsLoading(false)
+    }
+  }
+
+  function displayResponseMessage(message: React.ReactNode) {
+    setResponseMessage(message)
+    setTimeout(() => {
+      setResponseMessage(<p></p>)
+    }, 5000)
+  }
+
+  const onChange = (imageList: ImageListType) => {
+    setImages(imageList)
+  }
+
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<IFormDataAddProduct>()
+
+  const onSubmit = (data: IFormDataAddProduct) => {
+    createProduct(images, data.title, data.subTitle, data.price, data.onStock)
+  }
+
   return (
-    <>
+    <form onSubmit={handleSubmit(onSubmit)}>
+      <ImageUploading multiple value={images} onChange={onChange} dataURLKey="data_url">
+        {({ imageList, onImageUpload, onImageRemoveAll, onImageUpdate, onImageRemove, isDragging, dragProps }) => (
+          <div className="w-full flex flex-col items-center justify-center gap-y-4 w-[50%]">
+            <Button
+              className={`${
+                isDraggingg && "fixed inset-0 z-[101] !bg-[rgba(0,0,0,0.6)]"
+              } image-upload w-full bg-transparent px-16 py-8 text-xl whitespace-nowrap transition-all duration-300`}
+              ref={dragZone}
+              onClick={onImageUpload}
+              {...dragProps}>
+              <h1 className="pointer-events-none select-none">
+                {isDragging ? "Drop files here" : "Click or Drop here"}
+              </h1>
+            </Button>
+            &nbsp;
+            {imageList.map((image, index) => (
+              <div
+                key={index}
+                className="flex flex-col gap-y-4 overflow-hidden rounded-xl border-[1px] border-solid border-border-color">
+                <Image
+                  className="aspect-video min-w-[320px] object-cover"
+                  src={image.data_url}
+                  alt="iamge"
+                  width={0}
+                  height={0}
+                />
+                <div className="mb-4 flex flex-row items-center justify-center gap-x-4 px-4">
+                  <Button className="w-1/2" onClick={() => onImageUpdate(index)}>
+                    Update
+                  </Button>
+                  <Button className="w-1/2" variant="danger-outline" onClick={() => onImageRemove(index)}>
+                    Remove
+                  </Button>
+                </div>
+              </div>
+            ))}
+            <div className="mb-4 flex w-full flex-row items-center justify-center gap-x-4">
+              <Button className="w-full" variant="danger-outline" onClick={void onImageRemoveAll}>
+                Remove all images
+              </Button>
+            </div>
+          </div>
+        )}
+      </ImageUploading>
       <ProductInput
         id="title"
         register={register}
@@ -47,9 +185,11 @@ export default function AddProductForm({ register, errors, responseMessage, isLo
         placeholder="Amount on stock"
       />
       <div className="text-center">{responseMessage}</div>
-      <Button className={`${isLoading && "opacity-50 cursor-default pointer-events-none"}`} disabled={isLoading}>
+      <Button
+        className={twMerge(`w-full mt-2`, isLoading && "opacity-50 cursor-default pointer-events-none")}
+        disabled={isLoading}>
         Create product
       </Button>
-    </>
+    </form>
   )
 }
